+++ config/ember-try.js
@@ -36,6 +36,15 @@ module.exports = {
       }
     },
     {
+      name: 'ember-lts-2.12',
+      npm: {
+        devDependencies: {
+          'ember-source': '~2.12.0',
+          'ember-native-dom-event-dispatcher': null
+        }
+      }
+    },
+    {
       name: 'ember-release',
       bower: {
         dependencies: {