+++ tests/helpers/module-for-acceptance.js
@@ -1,8 +1,7 @@
 import { module } from 'qunit';
+import { resolve } from 'rsvp';
 import startApp from '../helpers/start-app';
 import destroyApp from '../helpers/destroy-app';
-import RSVP from 'rsvp';
-const { Promise } = RSVP;
 
 export default function(name, options = {}) {
   module(name, {
@@ -16,7 +15,7 @@ export default function(name, options = {}) {
 
     afterEach() {
       let afterEach = options.afterEach && options.afterEach.apply(this, arguments);
-      return Promise.resolve(afterEach).then(() => destroyApp(this.application));
+      return resolve(afterEach).then(() => destroyApp(this.application));
     }
   });
 }