+++ tests/integration/components/constants.js
@@ -21,6 +21,19 @@ export const numbers = [
   'twenty'
 ];
 
+export const numerals = [
+  '2',
+  '3',
+  '5',
+  '7',
+  '11',
+  '13',
+  '17',
+  '19',
+  '23',
+  '853'
+];
+
 export const names = [
   'María',
   'Søren Larsen',