+++ tests/integration/components/power-select/customization-with-components-test.js
@@ -5,6 +5,7 @@ import { groupedNumbers } from '../constants';
 import { clickTrigger } from '../../../helpers/ember-power-select';
 import { find, findAll, click } from 'ember-native-dom-helpers';
 import { get } from '@ember/object';
+import Component from '@ember/component';
 import { getOwner } from '@ember/application';
 import { isPresent } from '@ember/utils';
 
@@ -27,7 +28,20 @@ test('selected option can be customized using triggerComponent', function(assert
   assert.notOk(find('.ember-power-select-status-icon'), 'The provided trigger component is not rendered');
   assert.ok(find('.ember-power-select-trigger .icon-flag'), 'The custom flag appears.');
   assert.equal(find('.ember-power-select-trigger').textContent.trim(), 'Spain', 'With the country name as the text.');
+});
 
+test('triggerComponent receives loading message', function(assert) {
+  assert.expect(1);
+
+  this.register('component:custom-trigger-component', Component.extend({
+    layout: hbs`<div class="custom-trigger-component">{{loadingMessage}}</div>`
+  }));
+  this.render(hbs`
+    {{#power-select loadingMessage="hmmmm paella" selected=country triggerComponent="custom-trigger-component" onchange=(action (mut foo)) as |country|}}
+      {{option}}
+    {{/power-select}}
+  `);
+  assert.equal(find('.custom-trigger-component').textContent, 'hmmmm paella', 'The loading message is passed to the trigger component');
 });
 
 test('selected item option can be customized using selectedItemComponent', function(assert) {
@@ -84,19 +98,29 @@ test('the `optionsComponent` receives the `extra` hash', function(assert) {
 });
 
 test('the content before the list can be customized passing `beforeOptionsComponent`', function(assert) {
-  assert.expect(2);
+  assert.expect(4);
 
   this.countries = countries;
   this.country = countries[1]; // Spain
 
   this.render(hbs`
-    {{#power-select options=countries selected=country beforeOptionsComponent="custom-before-options" onchange=(action (mut foo)) as |country|}}
+    {{#power-select
+      options=countries
+      selected=country
+      beforeOptionsComponent="custom-before-options"
+      placeholder="inception"
+      placeholderComponent=(component "power-select/placeholder")
+      onchange=(action (mut foo)) as |country|}}
       {{country.name}}
     {{/power-select}}
   `);
 
   clickTrigger();
-  assert.ok(find('.ember-power-select-dropdown #custom-before-options-p-tag'), 'The custom component is rendered instead of the usual search bar');
+
+  let element = find('.ember-power-select-dropdown #custom-before-options-p-tag');
+  assert.ok(element, 'The custom component is rendered instead of the usual search bar');
+  assert.equal(element.innerText, 'inception', 'The placeholder attribute is passed through.');
+  assert.ok(find('.ember-power-select-dropdown .ember-power-select-placeholder'), 'The placeholder component is passed through.');
   assert.notOk(find('.ember-power-select-search-input'), 'The search input is not visible');
 });